Implement color-change inside another function
The problem was, that once a button was clicked you couldn't use it
again without reloading the page or creating a new gird. Now when a
buttons gets clicked, for example the colorful one, then the event-
listener for the black buttons gets removed until the black button is
clicked, which adds the eventlistener again and removes the colorful.
This is implemented using a boolean variable to check which color is
currently used